[Patchset 13] Change SHM to be handled per producer
The size of the shm and the page size can now be specified per producer,
through the config file.
The shared memory is now created at the same time as the first data source,
instead of when the connection is first established.
As a side effect, the SharedMemoryABI had to be refactored to allow an
default constructor.

TODO: tests, implement teardown of shm


Bug:74098471
Bug:69708583

+++ src/tracing/core/service_impl.cc
@@ -567,8 +567,8 @@ void ServiceImpl::CreateDataSourceInstance(
 
     producer->page_size_kb_ = first_producer_config.page_size_kb();
 
-    size_t shm_size =
-        std::min(first_producer_config.shm_size_kb(), kMaxShmSize);
+    size_t shm_size = std::min(
+        static_cast<size_t>(first_producer_config.shm_size_kb()), kMaxShmSize);
     if (shm_size % base::kPageSize || shm_size < base::kPageSize)
       shm_size = std::min(shared_memory_size_hint_bytes_, kMaxShmSize);
     if (shm_size % base::kPageSize || shm_size < base::kPageSize)
